backend/requirements.txt
tensorflow==2.19.0
numpy==1.26.1
pillow==10.0.0
fastapi==0.118.0
uvicorn==0.23.2
python-multipart==0.0.6
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.3.2
python-dotenv==1.0.0
pillow==10.1.0
werkzeug==3.0.1